pass proxy into collector instead of agent + boto client recreation fix
